Reduce opacity effect of scaled arrow label, update font in export, adds command+enter to finish text
