+++ packages/tldraw/src/state/TldrawApp.ts
@@ -1820,7 +1820,7 @@ export class TldrawApp extends StateManager<TDSnapshot> {
     // Embed our custom fonts
     const defs = document.createElementNS('http://www.w3.org/2000/svg', 'defs')
     const style = document.createElementNS('http://www.w3.org/2000/svg', 'style')
-    style.textContent = `@import url('https://fonts.googleapis.com/css2?family=Caveat+Brush&family=Source+Code+Pro&family=Source+Sans+Pro&family=Source+Serif+Pro&display=block');`
+    style.textContent = `@import url('https://fonts.googleapis.com/css2?family=Caveat+Brush&family=Source+Code+Pro&family=Source+Sans+Pro&family=Crimson+Pro&display=block');`
     defs.appendChild(style)
     svg.appendChild(defs)
     // Get the shapes in order
